snap-patch-backend/requirements.txt
 # Core
 fastapi==0.115.12
 uvicorn==0.34.2
 starlette==0.46.2
 python-multipart==0.0.20
 python-dotenv==1.0.1
 httpx==0.28.1
 httpcore==1.0.8
 anyio==4.9.0

 # AI / CV
 tensorflow==2.16.1
 tensorflow-hub==0.16.1
 mediapipe==0.10.21
 opencv-python==4.11.0.86
 numpy==1.26.4
 scipy==1.13.1
 Pillow==11.2.1
 absl-py==2.2.2
 trimesh==4.3.2          

 # Utils
 packaging==25.0
 typing_extensions==4.13.2
 typing-inspection==0.4.0

 # Testing
 pytest==8.3.5
 pluggy==1.5.0
 iniconfig==2.1.0
 exceptiongroup==1.2.2
 grpcio-status<1.35

# Auth0 security
python-jose[cryptography]
fastapi-auth0

# MongoDB Atlas
motor
pymongo[srv]

# Gemini Q&A
requests
